pyproject.toml
[project]
name = "newtimestep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy>=2.2.0",
    "pygame-ce>=2.5.2",
    "pyqt6>=6.8.0",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "ruff>=0.8.3",
]

[tool.mypy]
strict = true
show_column_numbers = true

[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
extend-select = ["I"]
ignore = ["E741"]
fixable = ["I"]

[tool.pytest.ini_options]
testpaths = "tests"
pythonpath = "."


